requirements.txt
langchain
langchain-openai
langchain-community
langchain-anthropic
pydantic
python-dotenv
langgraph
faiss-cpu
sentence-transformers
kafka-python
minio
qdrant-client